Move back into common/upstream
https://github.com/envoyproxy/envoy/pull/2960#issuecomment-377988994
