+++ source/common/upstream/BUILD
@@ -93,6 +93,35 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "health_checker_base_lib",
+    srcs = ["health_checker_base_impl.cc"],
+    hdrs = ["health_checker_base_impl.h"],
+    deps = [
+        "//include/envoy/upstream:health_checker_interface",
+        "//source/common/router:router_lib",
+        "@envoy_api//envoy/api/v2/core:health_check_cc",
+    ],
+)
+
+envoy_cc_library(
+    name = "health_checker_lib",
+    srcs = ["health_checker_impl.cc"],
+    hdrs = ["health_checker_impl.h"],
+    external_deps = [
+        "grpc_health_proto",
+    ],
+    deps = [
+        ":health_checker_base_lib",
+        "//source/common/grpc:codec_lib",
+        "//source/common/http:codec_client_lib",
+        "//source/common/upstream:host_utility_lib",
+        # TODO(mattklein123): The following should not be required once we allow pluggable health
+        # checkers.
+        "//source/extensions/filters/network/redis_proxy:conn_pool_lib",
+    ],
+)
+
+envoy_cc_library(
     name = "host_utility_lib",
     srcs = ["host_utility.cc"],
     hdrs = ["host_utility.h"],
@@ -294,6 +323,7 @@ envoy_cc_library(
     srcs = ["upstream_impl.cc"],
     deps = [
         ":eds_lib",
+        ":health_checker_lib",
         ":logical_dns_cluster_lib",
         ":original_dst_cluster_lib",
         ":upstream_includes",
@@ -313,7 +343,6 @@ envoy_cc_library(
         "//source/common/network:utility_lib",
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
-        "//source/common/upstream/health_checker",
         "@envoy_api//envoy/api/v2/core:base_cc",
     ],
 )
